Implement workaround for issue #14
Older Android SDKs don't implement MIME type "text/markdown". This leads to markdown files not being available for selection.

This workaround checks if "text/markdown" is a known MIME type. Otherwise it will fall back to the broader "*/*".

Possible issue with this workaround: Instead of not being able to select files, it is now possible to select any file regardless of file type.
This only affects devices, where "text/markdown" is not known.